00_openai_agents/01_hello_agent/sync/hello_agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.1",
]
